Problem: Options type is a new integer type
That exposes all operators and functions applicable to integers. That's
not desirable.

Solution: wrap it into a record

This removes the need to override operators.

While at it, simplified contracts. Though now it seems it requires me to
go through some of the specific provers to prove it (like z3). Not sure
why other provers can't prove it yet.